+++ datas/photos_array.php
@@ -1,7 +1,162 @@
-<?php 
+<?php
+
+//! IMPORTS  ###########################################################
+
+require_once($prefix_to_root_folder."models/media/Photo.php");
+require_once($prefix_to_root_folder."models/media/Album.php"); // class Album need class Photo to be declared first !
+require_once($prefix_to_root_folder."models/media/Gallery.php"); // class Album need class Photo to be declared first !
+require_once($prefix_to_root_folder."debug/debug_functions.php");
+
+
+use christophegourmand\debug as debug;
+
+use models\media\Photo as Photo;
+use models\media\Album as Album;
+use models\media\Gallery as Gallery;
+
+
+
+//! METHOD USING CLASSES  ##############################################
+
+//* CREATION ALBUM 1  **********************************************
+
+
+// METHOD USING CLASSES
+// #####################################################################
+
+
+// CREATION ALBUM 1
+// --------------------------------------------------------------
+
+
+/* $photo27 = new Photo( "alinea_00027_groupe.jpg", "les membres du groupe Alinéa",  "landscape"); */
+// showInHtml($photo27 , "photo27");
+
+/* $photo16 = new Photo( "alinea_00016_thierry-guitare.jpg", "Thierry à la guitare", "landscape"); */
+// showInHtml($photo16 , "photo16");
+
+$album1 = new Album(
+    "2020.06.24",
+    "/assets/img/photos/from-CoppaStudio/2020.06.24/Color/SD/thumbnails/",
+    "/assets/img/photos/from-CoppaStudio/2020.06.24/Color/SD/"
+);
+/*
+$album1->addPhoto($photo27);
+$album1->addPhoto($photo16);
+ */
+
+$album1->addPhoto( new Photo("alinea_00027_groupe.jpg", "les membres du groupe Alinéa", "landscape") );
+$album1->addPhoto( new Photo("alinea_00016_thierry-guitare.jpg", "Thierry à la guitare", "landscape") );
+$album1->addPhoto( new Photo("alinea_00019_instrument_cajon.jpg", "Instrument Cajon","landscape") );
+$album1->addPhoto( new Photo("alinea_00020_instrument_cajon.jpg", "Instrument Cajon", "landscape") );
+$album1->addPhoto( new Photo("alinea_00021_nicole.jpg", "Nicole", "landscape") );
+$album1->addPhoto( new Photo("alinea_00022_nico-tatouage.jpg", "Nico et son tatouage", "landscape") );
+$album1->addPhoto( new Photo("alinea_00024_instrument_steel-tongue-drum.jpg", "instrument: steel tongue drum", "landscape") );
+$album1->addPhoto( new Photo("alinea_00028_groupe.jpg", "Membres du groupe Alinéa", "landscape") );
+$album1->addPhoto( new Photo("alinea_00029_groupe.jpg", "Membres du groupe Alinéa", "landscape") );
+$album1->addPhoto( new Photo("alinea_00030_groupe.jpg", "Membres du groupe Alinéa", "landscape") );
+$album1->addPhoto( new Photo("alinea_00032_instrument_steel-tongue-drum.jpg", "instrument: steel tongue drum", "portrait") );
+
+$album1->addPhoto( new Photo("alinea_00034_instrument_guitare-thierry.jpg", "Instrument: guitare", "landscape") );
+// $album1->addPhoto( new Photo("alinea_00035_instrument_guitare-thierry.jpg", "Instrument: guitare", "landscape") );
+$album1->addPhoto( new Photo("alinea_00036_thierry.jpg", "Thierry chante", "landscape") );
+$album1->addPhoto( new Photo("alinea_00038_instrument_steel-tongue-drum.jpg", "instrument: steel tongue drum", "landscape") );
+$album1->addPhoto( new Photo("alinea_00039_arbre-raylight.jpg", "soleil dans les arbres", "landscape") );
+$album1->addPhoto( new Photo("alinea_00040_thierry-play-guitare.jpg", "Thierry joue de la guitare", "landscape") );
+$album1->addPhoto( new Photo("alinea_00041_nicole.jpg", "Nicole", "landscape") );
+$album1->addPhoto( new Photo("alinea_00042_instrument_piano.jpg", "instrument: piano", "landscape") );
+$album1->addPhoto( new Photo("alinea_00043_instrument_steel-tongue-drum.jpg", "instrument: steel tongue drum", "landscape") );
+$album1->addPhoto( new Photo("alinea_00044_instrument_claves.jpg", "instrument: claves", "landscape") );
+$album1->addPhoto( new Photo("alinea_00045_instrument_claves.jpg", "instrument: claves", "landscape") );
+$album1->addPhoto( new Photo("alinea_00048_instrument_tambourin.jpg", "instrument: tambourin", "landscape") );
+$album1->addPhoto( new Photo("alinea_00049_instrument_guitare-thierry.jpg", "instrument: accord de guitare", "landscape") );
+$album1->addPhoto( new Photo("alinea_00050_instrument_guitare-thierry.jpg", "instrument: chevalet de guitare", "landscape") );
+
+// debug\showInHtml($album1 , "album1"); // 🟥 DEBUG 
+
+
+
+// CREATION ALBUM 2
+// --------------------------------------------------------------
+
+$album2 = new Album(
+    "2020.10.24",
+    "/assets/img/photos/from-facebook/2020.10.23/thumbnails/",
+    "/assets/img/photos/from-facebook/2020.10.23/"
+);
+
+
+$album2->addPhoto( new Photo("alinea_00069_nicole.jpg", "Nicole, souriante.", "landscape") );
+$album2->addPhoto( new Photo("alinea_00070_thierry-play-guitare.jpg", "Thierry sur sa guitare électrique", "landscape") );
+$album2->addPhoto( new Photo("alinea_00071_nico-play-drums.jpg", "Nico, en plein riff de batterie", "landscape") );
+// $album2->addPhoto( new Photo("alinea_00072_romain-play-bass-bw.jpg", "Romain à la basse", "landscape") );
+$album2->addPhoto( new Photo("alinea_00073_romain-play-bass-color.jpg", "Romain à la basse", "landscape") );
+// $album2->addPhoto( new Photo("alinea_00074_simone-play-piano.jpg", "Simone au piano", "landscape") );
+// $album2->addPhoto( new Photo("alinea_00075_simone-play-piano-bw.jpg", "Simone au piano", "landscape") );
+$album2->addPhoto( new Photo("alinea_00076_thierry-play-guitare.jpg", "Thierry à la guitare", "landscape") );
+$album2->addPhoto( new Photo("alinea_00077_nico-play-drums.jpg", "Nico à la batterie", "landscape") );
+
+// debug\showInHtml($album2 , "album2");// 🟥 DEBUG
+
+// CREATION ALBUM 3
+// --------------------------------------------------------------
+$album3 = new Album(
+    "2021.01.12 Répétition",
+    "/assets/img/photos/from-facebook/2021.01.12/thumbnails/",
+    "/assets/img/photos/from-facebook/2021.01.12/"
+);
+
+$album3->addPhoto( new Photo("alinea_00078_romain-play-saxo.jpeg", "Romain joue du saxo", "portrait") );
+$album3->addPhoto( new Photo("alinea_00079_nicole-sing.jpeg", "Nicole au chant", "portrait") );
+$album3->addPhoto( new Photo("alinea_00081_instrument_piano.jpeg", "Piano de Simone", "portrait") );
+$album3->addPhoto( new Photo("alinea_00082_instrument_drums.jpeg", "nouvelle batterie de Nico", "portrait") );
+$album3->addPhoto( new Photo("alinea_00084_nico-with-headphones.jpeg", "Nico écoute au casque", "portrait") );
+$album3->addPhoto( new Photo("alinea_00085_thierry-play-guitar.jpeg", "Thierry sur sa guitare SG", "square") );
+$album3->addPhoto( new Photo("alinea_00086_romain-play-drums.jpeg", "Romain à la batterie", "portrait") );
+$album3->addPhoto( new Photo("alinea_00080_instrument_drums.jpeg", "La nouvelle batterie de Nico, rayonnante", "landscape") );
+
+// CREATION ALBUM 4
+// --------------------------------------------------------------
+$album4 = new Album(
+    "2021.07.08 Répétition",
+    "/assets/img/photos/from-facebook/2021.07.08/thumbnails/",
+    "/assets/img/photos/from-facebook/2021.07.08/"
+);
+
+$album4->addPhoto( new Photo("alinea_00087_romain-playing-bass-and-nico-playing-drums.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00088_thierry-guitar-nicole-sing.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00089_romain-bass.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00090_romain-bass-blackwhite.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00091_thierry-playing-guitar-and-nico-playing-drums.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00092_nicole-singing.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00093_instruments-cables.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00094_nicole-smiling.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00095_thierry_smiling.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00096_thierry-and-nico-playing-guitar-and-drums.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00097_nicole_singing.jpg", "", "portrait") );
+// $album4->addPhoto( new Photo("alinea_00098_thierry-playing-guitar-and-nicole-singing.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00099_thierry-playing-guitar-and-nicole-singing.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00100_instruments-cables.jpg", "", "landscape") );
+$album4->addPhoto( new Photo("alinea_00101_nicole_singing.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00102_romain-playing-bass-and-nico-playing-drums.jpg", "", "portrait") );
+$album4->addPhoto( new Photo("alinea_00103_nicole_singing.jpg", "", "portrait") );
+
+
+// ADD ALBUM 1 AND 2 INTO GALLERY
+// --------------------------------------------------------------
+
+$gallery1 = new Gallery();
+$gallery1->addAlbums($album1, $album2, $album3, $album4);
+
+// debug\showInHtml($gallery1 , "gallery1");// 🟥 DEBUG 
+
+
+// OLD METHOD (using arrays)
+// #####################################################################
+
 
 $photoAlbums_indArray = [
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     [
         "album_name"=> "2020.06.24",
         "path_to_thumbnails"=>"/assets/img/photos/from-CoppaStudio/2020.06.24/Color/SD/thumbnails/",
@@ -45,7 +200,7 @@ $photoAlbums_indArray = [
             ["filename"=>"alinea_00050_instrument_guitare-thierry.jpg", "description"=>"instrument: chevalet de guitare"]
         ]
     ],
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     [
         "album_name"=> "2020.10.24",
         "path_to_thumbnails"=>"/assets/img/photos/from-facebook/2020.10.23/thumbnails/",
@@ -62,7 +217,7 @@ $photoAlbums_indArray = [
             ["filename"=>"alinea_00077_nico-play-drums.jpg", "description"=>"Nico à la batterie"]
         ]
     ],
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     [
         "album_name"=> "2021.01.12 Répétition",
         "path_to_thumbnails"=>"/assets/img/photos/from-facebook/2021.01.12/thumbnails/",
@@ -79,7 +234,7 @@ $photoAlbums_indArray = [
             ["filename"=>"alinea_00080_instrument_drums.jpeg", "description"=>"La nouvelle batterie de Nico, rayonnante"]
         ]
     ],
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     [
         "album_name"=> "2021.07.08 Répétition",
         "path_to_thumbnails"=>"/assets/img/photos/from-facebook/2021.07.08/thumbnails/",
@@ -105,7 +260,7 @@ $photoAlbums_indArray = [
             ["filename"=>"alinea_00103_nicole_singing.jpg", "description"=>""]
         ]
     ]
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     // [
         // "album_name"=> "",
         // "path_to_thumbnails"=>"",
@@ -130,7 +285,7 @@ $photoAlbums_indArray = [
             // ["filename"=>"", "description"=>""],
         // ]
     // ],
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     // [
         // "album_name"=> "",
         // "path_to_thumbnails"=>"",
@@ -155,7 +310,7 @@ $photoAlbums_indArray = [
             // ["filename"=>"", "description"=>""],
         // ]
     // ],
-    // ############################################################## NEW ALBUM
+    // - - - - - - - - - - - - - - - - - - - - - - - - - NEW ALBUM
     // [
         // "album_name"=> "",
         // "path_to_thumbnails"=>"",