+++ views/pages/page_photos.php
@@ -29,6 +29,10 @@
         <?php include($prefix_to_root_folder."views/pages/".$page_name."/page-content.php"); ?>
         
         <?php include($prefix_to_root_folder."views/common/footer.php"); ?>
+
         <?php require_once($prefix_to_root_folder."views/common/load_js_scripts.php"); ?>
+
+        <!-- custom script, only for this page -->
+        <script src="<?php echo($prefix_to_root_folder.'assets/js/animationOnPhotos.js')?>"></script>
     </body>
 </html>