+++ views/pages/page_photos_big/page-content.php
@@ -1,29 +1,31 @@
 <main class="">
-    <?php include($prefix_to_root_folder.'datas/photos_array.php'); ?>
-
     <?php
-        // on page_photos , when user click a photo thumbnail, the url pass 2 variables that we collect here : 
-        $album_id = intval( $_GET['album_id'] );
-        $photo_id = intval( $_GET['photo_id'] );
+        include($prefix_to_root_folder.'datas/photos_array.php');
+        
+        require_once($prefix_to_root_folder."debug/debug_functions.php");
+        use christophegourmand\debug as debug;
 
-        $photoAlbum_albumName = $photoAlbums_indArray[$album_id]['album_name'];
-        $photoAlbum_path_to_thumbmails = $photoAlbums_indArray[$album_id]['path_to_thumbnails'];
-        $photoAlbum_path_to_originals = $photoAlbums_indArray[$album_id]['path_to_originals'];
-        $photos_of_album_indArr = $photoAlbums_indArray[$album_id]['photos_of_album_indArr'];
+        // on page_photos , when user click a photo thumbnail, the url pass 2 variables that we collect here : 
+        $album_id_from_url = intval( $_GET['album_id'] );
+        $photo_id_from_url = intval( $_GET['photo_id'] );
+        
+        $album_obj = $gallery1->getAlbum($album_id_from_url);
+        $photo_obj = $album_obj->getPhoto($photo_id_from_url);
+        // $photo_previous = $album_obj->getPhoto($photo_id_from_url - 1);
+        // $photo_next = $album_obj->getPhoto($photo_id_from_url + 1);
 
         // Definition of photo path (previous): 
-            if ( intval($photo_id) === 0 ) {
+            if ( intval($photo_id_from_url) === 0 ) {
                 $path_to_previous_photo = $prefix_to_root_folder.'views/pages/page_photos.php';
             } else {
-                $path_to_previous_photo = $prefix_to_root_folder.'views/pages/page_photos_big.php'.'?album_id='.$album_id.'&photo_id='.($photo_id-1);
+                $path_to_previous_photo = $prefix_to_root_folder.'views/pages/page_photos_big.php'.'?album_id='.$album_id_from_url.'&photo_id='.($photo_id_from_url - 1);
             }
 
         // Definition of photo path (next): 
-            if ( intval($photo_id) === intval(array_key_last($photos_of_album_indArr)) ) {
+            if ( intval($photo_id_from_url) >= $album_obj->getSize() - 1 ) {
                 $path_to_next_photo =  $prefix_to_root_folder.'views/pages/page_photos.php';
-                
             } else {
-                $path_to_next_photo =  $prefix_to_root_folder.'views/pages/page_photos_big.php'.'?album_id='.$album_id.'&photo_id='.($photo_id+1);
+                $path_to_next_photo =  $prefix_to_root_folder.'views/pages/page_photos_big.php'.'?album_id='.$album_id_from_url.'&photo_id='.($photo_id_from_url + 1);
             }
     ?>
 
@@ -33,7 +35,7 @@
         <a class="gallery-img-btn img-btn--back" href="<?= $prefix_to_root_folder.'views/pages/page_photos.php';?>">
             <i class="fas fa-th"></i>
         </a>
-        <h3 class="gallery-title"><?=$photoAlbum_albumName; ?></h3>
+        <h3 class="gallery-title"><?= $album_obj->getTitle(); ?></h3>
         <div class="gallery-body-big">
             <figure class="gallery-img-container">
                 <a class="gallery-img-btn img-btn--previous" href="<?= $path_to_previous_photo; ?>">
@@ -41,13 +43,17 @@
                 </a>
 
                 
-                <img class="gallery-img-big" src="<?=$photoAlbum_path_to_originals.$photos_of_album_indArr[$photo_id]['filename']?>" alt="<?= $photos_of_album_indArr[$photo_id]['description']; ?>">
+                <img 
+                    class="gallery-img-big" 
+                    src="<?= $album_obj->getPath_to_originals().$photo_obj->getFilename()?>" 
+                    alt="<?= $photo_obj->getDescription(); ?>"
+                >
                 
                 <a class="gallery-img-btn img-btn--next" href="<?= $path_to_next_photo; ?>">
                     <i class="fas fa-chevron-right"></i>
                 </a>
 
-                <figcaption class="gallery-img-description"><?= $photos_of_album_indArr[$photo_id]['description'];?></figcaption>
+                <figcaption class="gallery-img-description"><?= $photo_obj->getDescription(); ?></figcaption>
             </figure>
         </div>
     </div>